formationBackend/requirement.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.7
defusedxml==0.8.0rc2
Django==4.1.13
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
djoser==2.2.2
dnspython==2.6.1
idna==3.7
oauthlib==3.2.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pymongo==4.7.2
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.2.4
tzdata==2024.1
urllib3==2.2.1
